refactor(types): group channel types under common/types/channel
